+++ Scoreboard.hpp
@@ -24,9 +24,9 @@ public:
 
       // Conditionals to place the scores in the right location based on player number
       if (i == 0)
-	playerScore[i].setPosition(window.getSize().x / 4, 10);
+	playerScore[i].setPosition(float(window.getSize().x / 4), 10.0f);
       else if (i == 1)
-	playerScore[i].setPosition(window.getSize().x - window.getSize().x / 4, 10);
+	playerScore[i].setPosition(float(window.getSize().x - window.getSize().x / 4), 10.0f);
       
     }
   }
